Removed old ScrobbleDto and moved 2 enums to dedicated files since they are now core to the implementation.
