+++ Kavita.API/Repositories/IScrobbleRepository.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Kavita.Common.Helpers;
 using Kavita.Models.DTOs.Scrobbling;
 using Kavita.Models.Entities.Enums;
+using Kavita.Models.Entities.Enums.KavitaPlus;
 using Kavita.Models.Entities.History;
 using Kavita.Models.Entities.Scrobble;
