+++ Kavita.Models/DTOs/KavitaPlus/ExternalMetadata/MatchRequestV3Dto.cs
@@ -1,8 +1,10 @@
 using System.Collections.Generic;
 using Kavita.Models.DTOs.Scrobbling;
 using Kavita.Models.Entities.Enums;
+using Kavita.Models.Entities.Enums.KavitaPlus;
 
 namespace Kavita.Models.DTOs.KavitaPlus.ExternalMetadata;
+#nullable enable
 
 public sealed record MatchRequestV3Dto: MetadataRequest
 {
@@ -11,6 +13,5 @@ public sealed record MatchRequestV3Dto: MetadataRequest
     public List<string> AlternativeNames { get; set; } = [];
     public int? Year { get; set; }
     public string? Query { get; set; }
-
     public PlusMediaFormat Format { get; set; }
 }
