+++ Kavita.Models/DTOs/KavitaPlus/ExternalMetadata/SeriesDetailRequestV3Dto.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using Kavita.Models.DTOs.Scrobbling;
 using Kavita.Models.Entities.Enums;
+using Kavita.Models.Entities.Enums.KavitaPlus;
 
 namespace Kavita.Models.DTOs.KavitaPlus.ExternalMetadata;
