+++ Kavita.Models/DTOs/KavitaPlus/Metadata/ExternalSeriesDetailDto.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using Kavita.Models.DTOs.Recommendation;
 using Kavita.Models.DTOs.Scrobbling;
 using Kavita.Models.Entities.Enums;
+using Kavita.Models.Entities.Enums.KavitaPlus;
 
 namespace Kavita.Models.DTOs.KavitaPlus.Metadata;
