+++ Kavita.Models/DTOs/KavitaPlus/Metadata/SeriesRelationship.cs
@@ -1,6 +1,7 @@
 ﻿#nullable enable
 using Kavita.Models.DTOs.Scrobbling;
 using Kavita.Models.Entities.Enums;
+using Kavita.Models.Entities.Enums.KavitaPlus;
 
 namespace Kavita.Models.DTOs.KavitaPlus.Metadata;
