+++ Kavita.Models/DTOs/Metadata/Matching/MatchSeriesInfoDto.cs
@@ -1,5 +1,6 @@
 ﻿using Kavita.Models.DTOs.Scrobbling;
 using Kavita.Models.Entities.Enums;
+using Kavita.Models.Entities.Enums.KavitaPlus;
 
 namespace Kavita.Models.DTOs.Metadata.Matching;
