+++ Kavita.Models/DTOs/Scrobbling/ScrobbleEventDto.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Kavita.Models.Entities.Enums;
+using Kavita.Models.Entities.Enums.KavitaPlus;
 using Kavita.Models.Entities.Enums.UserPreferences;
 
 namespace Kavita.Models.DTOs.Scrobbling;
